+++ .idea/modules.xml
@@ -2,9 +2,11 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/Refereeing.iml" filepath="$PROJECT_DIR$/Refereeing.iml" />
-      <module fileurl="file://$PROJECT_DIR$/Slav-master.iml" filepath="$PROJECT_DIR$/Slav-master.iml" />
+      <module fileurl="file://C:\Users\Yaroslav\Slav\AndroidStudioProject\Reffery\danceReffery\Refereeing.iml" filepath="C:\Users\Yaroslav\Slav\AndroidStudioProject\Reffery\danceReffery\Refereeing.iml" />
+      <module fileurl="file://C:\Users\Yaroslav\Slav\AndroidStudioProject\Reffery\danceReffery\Slav-master.iml" filepath="C:\Users\Yaroslav\Slav\AndroidStudioProject\Reffery\danceReffery\Slav-master.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://C:\Users\Yaroslav\Slav\AndroidStudioProject\Reffery\danceReffery\app\app.iml" filepath="C:\Users\Yaroslav\Slav\AndroidStudioProject\Reffery\danceReffery\app\app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/danceReffery.iml" filepath="$PROJECT_DIR$/danceReffery.iml" />
     </modules>
   </component>
 </project>
