+++ app/src/main/java/ru/startandroid/refereeing/activity/dance/QuarterfinalTg.java
@@ -1,5 +1,6 @@
-package ru.startandroid.refereeing.activity;
+package ru.startandroid.refereeing.activity.dance;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
@@ -15,11 +16,11 @@ import ru.startandroid.refereeing.R;
  * Created by Slav on 26.12.2016.
  */
 
-public class QuarterfinalQw extends AppCompatActivity {
+public class QuarterfinalTg extends AppCompatActivity {
     private TextView tvDanceName;
-    private Button btnEndDance;
-    final String JUDGENAME = "judgeName";
     private TextView judgeName;
+    final String JUDGENAME = "judgeName";
+    private Button btnEndDance;
     private GridView gl_event1;
     private GridView gl_event2;
     String[] couples1 = {"1", "2","3" ,"4" ,"5" ,"6" ,"7", "8", "9", "10", "11", "12"};
@@ -31,13 +32,21 @@ public class QuarterfinalQw extends AppCompatActivity {
         setContentView(R.layout.activity_quarterfinaldance);
 
         tvDanceName = (TextView)findViewById(R.id.tvDanceName);
-        tvDanceName.setText("Квікстеп");
-
-        btnEndDance = (Button)findViewById(R.id.btnEndDance);
-        btnEndDance.setText("Завершити категорію");
+        tvDanceName.setText("Танго");
 
         judgeName = (TextView)findViewById(R.id.judgeName);
         judgeName.setText(getIntent().getExtras().getString(JUDGENAME));
+
+        btnEndDance = (Button)findViewById(R.id.btnEndDance);
+        btnEndDance.setText("Наступний танець");
+        final Intent intent = new Intent(this, QuarterfinalQw.class);
+        btnEndDance.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                intent.putExtra(JUDGENAME, tvDanceName.getText());
+                startActivity(intent);
+            }
+        });
 
         gl_event1 = (GridView)findViewById(R.id.gl_event1);
         gl_event2 = (GridView)findViewById(R.id.gl_event2);
@@ -49,15 +58,12 @@ public class QuarterfinalQw extends AppCompatActivity {
         adjustGridView(gl_event1);
         adjustGridView(gl_event2);
     }
-
     private void adjustGridView(GridView gridView) {
         gridView.setNumColumns(6);
         gridView.setColumnWidth(80);
         gridView.setVerticalSpacing(5);
         gridView.setHorizontalSpacing(5);
     }
-
-
     public void onclick (View view){
         view.setBackgroundResource(R.drawable.rect_green);
     }
